pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegraph-api"
version = "1.0.0"
description = "A complete Python wrapper for the Telegraph API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Telegraph Python Contributors"}
]
maintainers = [
    {name = "Telegraph Python Contributors"}
]
keywords = [
    "telegraph",
    "telegraph-api",
    "telegra.ph",
    "api",
    "wrapper",
    "client",
    "publishing",
    "blogging"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.25.0",
]

[project.optional-dependencies]
async = [
    "aiohttp>=3.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
all = [
    "telegraph-api[async,dev]",
]

[project.urls]
Homepage = "https://github.com/telegraph-py/telegraph-api"
Documentation = "https://github.com/telegraph-py/telegraph-api#readme"
Repository = "https://github.com/telegraph-py/telegraph-api"
"Bug Tracker" = "https://github.com/telegraph-py/telegraph-api/issues"
Changelog = "https://github.com/telegraph-py/telegraph-api/releases"

[tool.setuptools]
packages = ["telegraph"]

[tool.setuptools.package-data]
telegraph = ["py.typed"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["telegraph"]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
